feat(wintertc): migrate URL from boa_runtime
Moves `url` into `boa_wintertc` and points `boa_runtime` at it. The only
addition to the module is a free `register` entry point wrapping
`Url::register`, matching the signature `boa_wintertc::register` expects.

The `url` crate moves with the module, so `boa_runtime`'s `url` feature now
just forwards to `boa_wintertc/url`. It stays a dev-dependency of
`boa_runtime` because the `fetch` e2e tests use it directly.
