+++ core/runtime/Cargo.toml
@@ -22,13 +22,13 @@ http = { workspace = true, optional = true }
 reqwest = { workspace = true, optional = true }
 rustc-hash = { workspace = true, features = ["std"] }
 serde_json = { workspace = true, optional = true }
-url = { workspace = true, optional = true }
 
 [dev-dependencies]
 indoc.workspace = true
 rstest.workspace = true
 textwrap.workspace = true
 temp-env.workspace = true
+url.workspace = true
 
 [lints]
 workspace = true
@@ -47,7 +47,7 @@ default = [
 ]
 
 all = ["default", "reqwest-blocking"]
-url = ["dep:url"]
+url = ["boa_wintertc/url"]
 fetch = [
     "dep:either",
     "dep:http",
